XP-264 break up grpc/coordinator.py into modules
