+++ xain_fl/coordinator/__init__.py
@@ -1,0 +1,8 @@
+"""Defines constants to be used throughout the module
+"""
+
+# TODO: move to `xain_fl.config.py`
+# once https://xainag.atlassian.net/browse/XP-265 is done
+_ONE_DAY_IN_SECONDS: int = 60 * 60 * 24
+HEARTBEAT_TIME: int = 10
+HEARTBEAT_TIMEOUT: int = 5
