+++ xain_fl/serve.py
@@ -1,0 +1,43 @@
+import threading
+import time
+from concurrent import futures
+
+import grpc
+
+from xain_fl.coordinator import _ONE_DAY_IN_SECONDS
+from xain_fl.coordinator.coordinator import Coordinator
+from xain_fl.coordinator.coordinator_grpc import CoordinatorGrpc
+from xain_fl.coordinator.heartbeat import monitor_heartbeats
+from xain_fl.cproto import coordinator_pb2_grpc
+from xain_fl.logger import get_logger
+
+logger = get_logger(__name__)
+
+
+def serve(coordinator: Coordinator, host: str = "[::]", port: int = 50051) -> None:
+    """Main method to start the gRPC service.
+
+    This methods just creates the :class:`~.Coordinator`, sets up all threading
+    events and threads and configures and starts the gRPC service.
+    """
+    terminate_event = threading.Event()
+    monitor_thread = threading.Thread(
+        target=monitor_heartbeats, args=(coordinator, terminate_event)
+    )
+
+    server = grpc.server(futures.ThreadPoolExecutor(max_workers=10))
+    coordinator_pb2_grpc.add_CoordinatorServicer_to_server(
+        CoordinatorGrpc(coordinator), server
+    )
+    server.add_insecure_port(f"{host}:{port}")
+    server.start()
+    monitor_thread.start()
+
+    logger.info("Coordinator waiting for connections...")
+
+    try:
+        while True:
+            time.sleep(_ONE_DAY_IN_SECONDS)
+    except KeyboardInterrupt:
+        terminate_event.set()
+        server.stop(0)
